split: ASocket -> ClientSocket, ServerSocket
